api/views.py
from rest_framework import generics

from .models import Post
from .serializers import PostSerializer

# import pygeoip


class PostList(generics.ListCreateAPIView):
    queryset = Post.objects.all()
    serializer_class = PostSerializer


class PostDetail(generics.RetrieveDestroyAPIView):
    queryset = Post.objects.all()
    serializer_class = PostSerializer
