requirements.txt
asgiref==3.5.2
black==22.3.0
click==8.1.3
Django==4.0.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
flake8==4.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pycodestyle==2.8.0
pyflakes==2.4.0
pygeoip==0.3.2
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.2.0
